Make traccar configuration more understandable
When I tried to setup a connection between traccar server and owntracks app, I found the documentation to be quite frustrating. After I had a working setup, I was able to understand the documentation, but from first reading I understood almost everything wrongly. Here I try to improve it, by changing the overall perspective of the text and make it more useful for general audience.